completed authentication

+++ src/types/product/product.resolvers.js
@@ -11,6 +11,9 @@ const productsTypeMatcher = {
 
 /** product */
 const product = (_, args, ctx) => {
+  if (!ctx.user) {
+    throw new AuthenticationError('not allowed')
+  }
   return Product.findById(args.id)
     .lean()
     .exec()
@@ -18,24 +21,37 @@ const product = (_, args, ctx) => {
 
 const newProduct = (_, args, ctx) => {
   // use this fake ID for createdBy for now until we talk auth
-  const createdBy = mongoose.Types.ObjectId()
+  if (!ctx.user || ctx.user.role !== roles.admin) {
+    throw new AuthenticationError('not allowed')
+  }
+  // const createdBy = mongoose.Types.ObjectId()
+  const createdBy = ctx.user._id
   return Product.create({ ...args.input, createdBy })
 }
 
 const products = (_, args, ctx) => {
+  if (!ctx.user) {
+    throw new AuthenticationError('not allowed')
+  }
   return Product.find({})
     .lean()
     .exec()
 }
 
 const updateProduct = (_, args, ctx) => {
+  if (!ctx.user || ctx.user.role !== roles.admin) {
+    throw new AuthenticationError('not allowed')
+  }
   const update = args.input
-  return Product.findByIdAndUpdate(args.id, update, { new: true })
+  return Product.findByIdAndUpdate(ctx.user.id, update, { new: true })
     .lean()
     .exec()
 }
 
 const removeProduct = (_, args, ctx) => {
+  if (!ctx.user || ctx.user.role !== roles.admin) {
+    throw new AuthenticationError('not allowed')
+  }
   return Product.findByIdAndRemove(args.id)
     .lean()
     .exec()
